+++ allennlp_models/vision/dataset_readers/vision_reader.py
@@ -122,6 +122,9 @@ class VisionReader(DatasetReader):
         # feature cache
         self.feature_cache_dir = feature_cache_dir
         self.coordinates_cache_dir = feature_cache_dir
+        self.class_probs_cache_dir = feature_cache_dir
+        self.class_labels_cache_dir = feature_cache_dir
+
         if feature_cache_dir:
             self.write_to_cache = write_to_cache
         else:
@@ -130,6 +133,8 @@ class VisionReader(DatasetReader):
             self.write_to_cache = True
         self._feature_cache_instance: Optional[MutableMapping[str, Tensor]] = None
         self._coordinates_cache_instance: Optional[MutableMapping[str, Tensor]] = None
+        self._class_probs_cache_instance: Optional[MutableMapping[str, Tensor]] = None
+        self._class_labels_cache_instance: Optional[MutableMapping[str, Tensor]] = None
 
         # image processors
         self.image_loader = None
@@ -234,9 +239,37 @@ class VisionReader(DatasetReader):
 
         return self._coordinates_cache_instance
 
+    @property
+    def _class_probs_cache(self) -> MutableMapping[str, Tensor]:
+        if self._class_probs_cache_instance is None:
+            if self.class_probs_cache_dir is None:
+                self._class_probs_cache_instance = {}
+            else:
+                os.makedirs(self.feature_cache_dir, exist_ok=True)  # type: ignore
+                self._class_probs_cache_instance = TensorCache(
+                    os.path.join(self.feature_cache_dir, "class_probs"),  # type: ignore
+                    read_only=not self.write_to_cache,
+                )
+
+        return self._class_probs_cache_instance
+
+    @property
+    def _class_labels_cache(self) -> MutableMapping[str, Tensor]:
+        if self._class_labels_cache_instance is None:
+            if self.class_labels_cache_dir is None:
+                self._class_labels_cache_instance = {}
+            else:
+                os.makedirs(self.feature_cache_dir, exist_ok=True)  # type: ignore
+                self._class_labels_cache_instance = TensorCache(
+                    os.path.join(self.feature_cache_dir, "class_labels"),  # type: ignore
+                    read_only=not self.write_to_cache,
+                )
+
+        return self._class_labels_cache_instance
+
     def _process_image_paths(
         self, image_paths: Iterable[str], *, use_cache: bool = True
-    ) -> Iterator[Tuple[Tensor, Tensor]]:
+    ) -> Iterator[Tuple[Tensor, Tensor, Optional[Tensor], Optional[Tensor]]]:
         """
         Processes the given image paths and returns featurized images.
 
@@ -258,7 +291,7 @@ class VisionReader(DatasetReader):
             "an image featurizer, and a region detector."
         )
 
-        batch: List[Union[str, Tuple[Tensor, Tensor]]] = []
+        batch: List[Union[str, Tuple[Tensor, Tensor, Optional[Tensor], Optional[Tensor]]]] = []
         unprocessed_paths: Set[str] = set()
 
         def yield_batch():
@@ -272,16 +305,44 @@ class VisionReader(DatasetReader):
                 detector_results = self.region_detector(images, sizes, featurized_images)
                 features = detector_results.features
                 coordinates = detector_results.boxes
+                class_probs = detector_results.class_probs
+                class_labels = detector_results.class_labels
 
             # store the processed results in memory, so we can complete the batch
-            paths_to_tensors = {path: (features[i], coordinates[i]) for i, path in enumerate(paths)}
+            paths_to_tensors = {}
+            for i, path in enumerate(paths):
+                if class_probs:
+                    class_probs_tensor = class_probs[i]
+                else:
+                    class_probs_tensor = None
+
+                if class_labels:
+                    class_labels_tensor = class_labels[i]
+                else:
+                    class_labels_tensor = None
+
+                paths_to_tensors[path] = (
+                    features[i],
+                    coordinates[i],
+                    class_probs_tensor,
+                    class_labels_tensor,
+                )
 
             # store the processed results in the cache
             if use_cache and self.write_to_cache:
-                for path, (features, coordinates) in paths_to_tensors.items():
+                for path, (
+                    features,
+                    coordinates,
+                    class_probs,
+                    class_labels,
+                ) in paths_to_tensors.items():
                     basename = os.path.basename(path)
                     self._feature_cache[basename] = features
                     self._coordinates_cache[basename] = coordinates
+                    if class_probs is not None:
+                        self._class_probs_cache[basename] = class_probs
+                    if class_labels is not None:
+                        self._class_labels_cache[basename] = class_labels
 
             # yield the batch
             for b in batch:
@@ -296,10 +357,12 @@ class VisionReader(DatasetReader):
                 if use_cache:
                     features: Tensor = self._feature_cache[basename]
                     coordinates: Tensor = self._coordinates_cache[basename]
+                    class_probs: Optional[Tensor] = self._class_probs_cache.get(basename)
+                    class_labels: Optional[Tensor] = self._class_labels_cache.get(basename)
                     if len(batch) <= 0:
-                        yield features, coordinates
+                        yield features, coordinates, class_probs, class_labels
                     else:
-                        batch.append((features, coordinates))
+                        batch.append((features, coordinates, class_probs, class_labels))
                 else:
                     # If we're not using the cache, we pretend we had a cache miss here.
                     raise KeyError